Move utils function (questionnaire -> splitting)
