+++ src/utils/lib/splitting.js
